Change column-drag-handle
It was set to 50% because I'm used to 30px headers. Seems it's supposed to be @header-height - 15px. And, now, so it is.

(Sadly, it doesn't play well with header heights much above 50px because of the drag sprite being on a spritesheet. You can get away with a little over, though.