Fix alignment issue with expand
When expand was turned on, the text went all funky and the bottom of the
header went weird. With this new setup, the blue expands to at least the
bottom of the button (based on header-height).

In the process, I ended up fixing the margin at the top of the blue-ness
as well. There's now a 1px border with the dark layout, like there is
around the rest of the top of the page, so that it aligns with the top
of the button. This is unnecessary on the light theme, since there is no
border to align with.